Add Config#default_gateway
The subscriptions gem requires a credit card in order to setup recurring
billing. This configuration method will attempt to take the value of an
instance variable if it has been configured, and fallback to the first
active gateway which supports credit cards.

If not such gateway exists it raises a friendly error informing the
developer how they can configure the default gateway in their app