Consolidated order handles out of stock error
If a particular line item is out of stock the consolidated order should
call the out of stock handler for the installment. This installment
should also be removed from the active installments the consolidated
installment is processing. This will avoid it having a completed order
associated to it which does not actually fulfill the failed installment