Handle no line items in cart
If there are no line items in the cart the order should not be
processed. It is possible this occured because of no available stock, in
which case the affected installments are already set to be reprocessed.
We do not raise an error and rollback the transaction so that the
changes to the affected installments are not lost and the record of the
attempt to process is persisted.

Any other uncaught errors will cause the whole transaction to be rolled
back.