Configurable subscription line item params
This is useful in the case where certain fields should not be allowed to
be modified by the user. This locks these attributes from being passed
in to the orders controller (or the api controller).

Ie if a store does not want to allow users to configure the number of
installments they will receive. Add this to an initializer:

SolidusSubscriptions::Config.subscription_line_item_attributes = [
  :quantity,
  :interval,
  :subscribable_id
]

This configuration also easily allows the gem to be customized to track
more information on the subcriptions line items if someone wanted to do
that