SubscriptionLineItem#dummy_line_item
Expose the dummy line item which is a representation of a line item
which would be added to a subscription order when the subscription was
processed.

It can be used to render the subcription_line_item in the cart. The
object is a frozen, unpersisted instance and cannot be saved