Update actionable_date when interval changes
If a subscription's line_item interval changes, we need to update the
actionable_date on the subscription in case the new interval means we
need to process the subscription immediately.

For example, if you have an interval of 3 months, and you change it to 1
month 2 months in, we should update the subscription to be actionable
immediately, as the new interval has passed since the last installment.