Fix case where actionable_date could be in past
If the subscription has no installments, and was created more than the
new interval ago, then the actionable_date could be set to the past. To
combat this, we refactor the method to now have two steps: find the
correct "base date" based upon whether we have installments or not, then
separately calculate the new interval and determine whether it's in the
past or not.