Only log failed installments
Previously, The Out of stock dispatcher was initialized with an empty
array and the system would log that it had failed to process <blank>
installments due to lack of stock.

The Failure dispatcher would be initalized with a blank array if the
order was successfully placed.

Both of the se dispatchers should only be activated if there is an
actially installment to justify notifying someone of this event