Raise error for invalid states
Do not allow scoping on invalid states.

Previously

SolidusSubscription::Subscription.in_processing_state(:mickey_mouse)

would return the entire collection of subscriptions. I think this is
rather unintuitive and I would rather get yelled at if I am trying to
filter by an unknown state.

To this effect. Raise an error if the state is not caught in the case
statement