Spec store credit subs orders
This is a sanity check spec to make sure store credits are applied to
subscription orders during the checkout.

This is solidus default behaviour, and while it may not be the desired
behaviour for every subscription implementation. Changeing this default
behaviour could be intensive. I am opting to wait for this to be a
feature requeset before implementing the ability to blacklist store
credit on subs orders