Cancel failed subscription orders
If a subscription order cannot be completed then it should be
canceled without notifying the user.

An arbitrary error can cause the process to fail prior to items being
added to the order. In this case we do not want to validate for line
items before canceling the order

This overrides the :ensure_line_items_present and :send_cancel methods
on a subscription orders.

Subscriptions orders can be cancelled from the cart with no line items
in them