Fix spec failures
Not really sure how this was correct before.

When the line item already existed on an order, the builder would
increment the quantity of the line item without saving it. in the specs,
a different instance of the line item would be tested resulting in the
change in quantity not being apparent to the spec