Use hash syntax for enums
This is much safer than the array syntax. the db indexes are explicitely
set on the model and are less likely to be accidentally changed.

These indexes are in the same order as the array was defined, no app
should notice a difference unless they have changed the order of this
array