Process user subs by address
Subscriptions are grouped together by user and and by shipping address.

A user has 3 subscriptions (A, B and C). A and B have the same shipping
address, but C will be shipped to a different location.

The processor will group and process installments for A and B together
and they will be fulfilled by the same order.

A second order will be created to fulfill C.