The dummy order uses best address available
If the subscription_line_item has an associated subscription, there is
likely a more valid shipping address than the shipping address of the
original order.

Create dummy orders by preferring the shipping address directly
associated with the subscription. The fallbacks for this a the shipping
address of the user associated with the subscription and the shipping
address of the origin order (if it exisits)