Replace max_installments with end_date
The max_installments concept of a specific number of times to processa
subscription was complicated and diffcult to understand. It was
originally meant to allow for subscriptions to be active only for a
given periode of time (ie, a 1 years magazine subscription).

To better model this behaviour, this column is being changed to
:end_date. A date after which the subscription will become inactive and
not be processed. If a subscription has no value for this field, it will
be processed until some action is taken to stop it (ex. cancelling the
subscription).