Remove some n+1 queries from processing
The processor looks up a number of users, subscriptions and installments
and passes them off to an Active Job.

Before we were doing a number of n+1 queries and relying on the rails
global id when passing objects to the consolidated installments.

This commit removes the greedy queries and passes a list of ids to the
active job object so that all affected objects can be looked up in a
single query