Add Subcription inputs to product page
In order to make the subscriptions gem a little more pluggable, I'm
including an override of the default frontend products page to input
form fields for subscriptions.

This should allow developers to simply "plug" the gem into their app and
see how it works. This is to make it easier to get a working example of
the subscriptions gem and encourage more feedback on the actual gem
funtionality