Authorize subscriptions admin sidebar link
Only show the admin sidebar link if the user is authorized to see it.

Before:
The link to the subscriptions area of the subscriptions administrative
area would be displayed on the admin login screen.
Clicking this link without authenticating would redirect the user to the
log in page

After:
The link to the subscriptions area of the admin is hidden until after
the user authenticates