Add ffaker dependency to gemspec
ffaker was removed as a runtime dependency of Solidus. As of right now,
this removal is only on Solidus master (but it is slated for v2.4).

Even though this extension does not use ffaker directly, Solidus
complains that its factories require the library and so the tests fail
without this.

PR removing ffaker from Solidus:
https://github.com/solidusio/solidus/pull/2163