Use Spree::UserClassHandle
This stops a deprecation in Rails. Rails 5.2 will raise an
ArgumentError if `class_name` is given a class. Instead Rails wants it
to be given a string.

Spree::UserClassHandle is a placeholder for the name of
Spree.user_class.

Check out `solidus/core/app/models/spree/user_class_handle.rb` for more
information.