Enable guest checkout for subscription orders
In the past, an existing and logged-in User account has been a
requirement for checking out with subscription items in the cart. In
order to allow customers to defer account creation, we can create the
User record at subscription generation time and allow the user to reset
the password later through the normal process. This should increase
conversion by giving users more choice in how they check out.

In the case where a user uses guest checkout but already has an account,
we associate the existing user with their order.

One possible improvement would be to use some existing facility for
looking up users by email. Ours is currently not case-insensitive.