Don't give cancelled subscription orders a `completed_at` value
By default, solidus_subscriptions sets the `completed_at` value on an
order even if payment fails. This causes the order to appear in the
admin's list of orders, which can be confusing. Instead, we skip that
step so that the orders don't appear with the other orders. We will need
to devise an alternate method of informing store owners / customers when
their subscription payments fail to process.