Prevent installment stacking
Recently, one of our customers reported that they observed a
subscription installment order on their account for qty. 2 of the
subscribed item, when they expected qty. 1. Investigation revealed that
their subscription (1 per month) had been without valid payment information for >1
month. Because of this, when they finally added valid payment info their
order covered 2 installments in effect "catching them up" for the missed
installments. This behavior is confusing for the customer, so we want to
change it slightly. Here, we skip creation of a new installment if the
subscription already has any actionable installments.