Allow setting subscription interval and end date via the API
Once a subscription is created, the interval and end date are not
read from the line items anymore, but rather from the subscription
itself. Therefore, users should be allowed to edit the interval and
end date on the subscription, and forbidden from editing it on the
line items, which will have no effect.

See issue [#114](https://github.com/solidusio-contrib/solidus_subscriptions/issues/114)
for more details about the matter.