Improve dummy order and dummy line item generation
Previously, `#dummy_line_item` would persist a new order to the DB on
every call. It now uses an order that is built on the fly and not
persisted to the DB.

Furthermore, the `Subscription` model now also exposes a `#dummy_order`
object that allows users to calculate the value of an entire subscription.