Fail silently if order cancellation transition fails
In Solidus 2.11, it isn't possible to cancel an order that hasn't been
completed anymore[1], which causes our failure dispatchers to fail when
they call `order.cancel!`.

To work around the issue, we're now only cancelling if the state machine
allows us to (i.e., on Solidus <= 2.10 and in customized state machines).
If the order cannot be cancelled, we skip the transition silently.

[1]: https://github.com/solidusio/solidus/pull/3542