Fix automatic cancellation being too aggressive
There was a bug in the `maximum_reprocessing_time` option that was
causing subscriptions to be cancelled by the system way before they
were actually due for cancellation.

Rather than looking at the date of the first failed payment, we were
looking at the date of the last successful payment. This means that, if
a customer was on a 4-week cycle and a payment failed, the system would
assume that payment had been failing for 4 weeks, where in fact it had
just started failing.

The new implementation looks at the date of the first failed charge after
the last successful charge, which is the behavior we expected.