Fix legacy unsubscribable subscriptions not being processed
If a customer subscribed to an item that is now unsubscribable, we don't
want to stop processing their subscription all of a sudden. Admins should
be able to decide whether they want to grandfather the customers out of
the subscription or keep them on the subscription indefinitely, while
preventing new customers from subscribing.