Remove time component from actionable_date columns
Given that the smallest interval_units is `day`, there's no reason for
actionable_date to be a datetime column. In fact, it can easily cause
subscriptions not to be processed, if the date on the column is equal
to the current date, but the time is in the future.

We should just turn the actionable_date columns into date columns, as
the name suggests, and assume that subscriptions will be processed once
per day at most, as is the case in the vast majority of apps.