Update cancel method on subscription to respect minimum date
Previously, the cancel method would check if the minimum
cancellation date minus the current day is in the future before
cancelling.

However, that adds one day to whatever minimum cancellation date
that you set - if you set one day, you would expect the user not
to be able to cancel on the day that the subscription ships, but
because the method checks if the date is in the future, the user
can't cancel the day of, or the day before.

This checks if the date is in the future OR if it's today, which
removes the added day from the minimum_cancellation_notice.

Also this sets the default to 0.days, as 1.day is pretty arbitrary

Lastly, this locks canonical-rails down to 0.2.9, because Solidus
still uses the `whitelisted_attributes` method, which was renamed
in 0.2.10