Update permissions check on user subscription tab
Previously, it would check your permissions on an ActiveRecord
association - which would in most non-admin cases be false, even
if they had permissions to manage subscriptions. This changes the
check to return a new instance of the users subscriptions, so
people with manage subscriptions permissions will still be able
to view this tab, while also allowing for you to conditionally
allow/deny permissions based on what users subscriptions they
are attempting to view.