Set promotion rules model paths similar to Solidus
This allows for naming the models in a way more consistent to Solidus
(which doesn't include "PromotionRule" in the name of the class).
Also, it easily allows for the promotion rules to be displayed with a
better name in the admin panel (which just uses the promotions model
name humanized).