Reload order during finalize action
If you decorate the subscription generator to change the order
in any way (for instance, to associate the order to the
subscription), that change is not persisted in the rest of the
finalize action, as the order has already been loaded by that
point.

This can cause massive headaches with validations, so to fix it
I've added a reload method before super, so we can ensure that
the order is up-to-date before continuing.

This also means that subscription_line_items interval is nil,
since activating the subscription sets it to nil. So in the spec
we're fetching the interval from the subscription instead.