Update the PDP subscription form
Fix the subscription PDP additional fields removing some strange
behaviur. Show the form only if the product or the one of the product
variants are subscribable.
Use the PDP quantity field to set the subscription quantity.
Add default values for the interval units fields.
Set the selected variant for the subscribable id.