Add the ability to pause subscriptions
Paused subscriptions will not be processed until a manual `#resume`,
or the `actionable_date` is reached. If the `actionable_date` is not
set, the subscription will be `paused` indefinitely.

If, and when, the `actionable_date` is reached, the `pause` column
for the subscription will be set to `false` in
`#advance actionable_date`