Run legacy promotion handler "cart" if legacy promo system present
In the legacy promotion system, there's a separation of tasks between
`order.recalculate` and `PromotionHandler::Cart`. `order.recalculate`
will only recalculate already existing promotion adjustments, but will
not check for any new promotions that might need to be applied. In the
new promotion system, both of these tasks are handled by
`order.recalculate`.

If the legacy promotion system is present, but not active, calling this
promotion handler will result in a single database call, which is an
acceptable price to pay I think.