Install generator: Don't fail if starter frontend not present
We're expecting the install generator of this gem to change things in
the starter frontend, but the generated Dummy app runs on the legacy
frontend. This fails, migrations don't run, and all tests fail.

This is not the cleanest solution: Ideally we'd have an install
generator for dummy apps that works with the starter frontend template,
but we're not there.